Chop chop chop. Swing Swing.
