+++ components/stsdk/stsdk_main.c
@@ -28,20 +28,30 @@
 #include <stdint.h>
 #include <inttypes.h>
 
-#include "stsdk_main.h"
-
+// esp includes
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
 #include "freertos/event_groups.h"
-
+#include "esp_system.h"
+#include "nvs_flash.h"
+#include "nvs.h"
+#include <lwip/netdb.h>
+#include "driver/uart.h"
 #include "esp_log.h"
+static const char *TAG = "STSDK";
 
+// AlarmDecoder includes
+#include "alarmdecoder_main.h"
+#include "ad2_utils.h"
+#include "ad2_settings.h"
+#include "ad2_uart_cli.h"
+
+#include "stsdk_main.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-static const char *TAG = "STSDK";
 
 iot_status_t g_iot_status = IOT_STATUS_IDLE;
 iot_stat_lv_t g_iot_stat_lv;
@@ -70,14 +80,127 @@ IOT_CAP_HANDLE *ota_cap_handle = NULL;
 IOT_CAP_HANDLE *healthCheck_cap_handle = NULL;
 IOT_CAP_HANDLE *refresh_cap_handle = NULL;
 
+
 #ifdef __cplusplus
 }
 #endif
 
-/*
- * Smartthings
- *
+/**
+ * Forget all STSDK state revert back to adoption mode.
  */
+static void _cli_cmd_cleanup_event(char *string)
+{
+    ESP_LOGI(TAG, "%s: clean-up data with reboot option", __func__);
+    st_conn_cleanup(ctx, true);
+}
+
+/**
+ * Configure the AD2IoT SmartThings device_info.json
+ *  field stored in NV
+ *
+ *  command: stserial <serialNumber>
+ * ex.
+ *   stserial AaBbCcDdEeFf...
+ */
+static void _cli_cmd_stserial_event(char *string)
+{
+    char arg[80];
+    if (ad2_copy_nth_arg(arg, string, sizeof(arg), 1) >= 0) {
+        nvs_handle my_handle;
+        esp_err_t err;
+        err = nvs_open_from_partition("stnv", "stdk", NVS_READWRITE, &my_handle);
+        if ( err != ESP_OK)
+            ESP_LOGE(TAG, "%s: nvs_open_from_partition err %d", __func__, err);
+        err = nvs_set_str(my_handle, "SerialNum", arg);
+        if ( err != ESP_OK)
+            ESP_LOGE(TAG, "%s: nvs_set_str err %d", __func__, err);
+        err = nvs_commit(my_handle);
+        nvs_close(my_handle);
+    } else {
+        printf("Missing <serialNumber>\n");
+    }
+}
+
+/**
+ * Configure the AD2IoT SmartThings device_info.json
+ *  field stored in NV
+ *
+ *  command: stpublickey <publicKey>
+ * ex.
+ *   stpublickey AaBbCcDdEeFf...
+ */
+static void _cli_cmd_stpublickey_event(char *string)
+{
+    char arg[80];
+    if (ad2_copy_nth_arg(arg, string, sizeof(arg), 1) >= 0) {
+        nvs_handle my_handle;
+        esp_err_t err;
+        err = nvs_open_from_partition("stnv", "stdk", NVS_READWRITE, &my_handle);
+        if ( err != ESP_OK)
+            ESP_LOGE(TAG, "%s: nvs_open_from_partition err %d", __func__, err);
+        err = nvs_set_str(my_handle, "PublicKey", arg);
+        if ( err != ESP_OK)
+            ESP_LOGE(TAG, "%s: nvs_set_str err %d", __func__, err);
+        err = nvs_commit(my_handle);
+        nvs_close(my_handle);
+    } else {
+        printf("Missing <publicKey>\n");
+    }
+}
+
+/**
+ * Configure the AD2IoT SmartThings device_info.json
+ *  field stored in NV
+ *
+ *  command: stprivatekey <privateKey>
+ * ex.
+ *   stprivatekey AaBbCcDdEeFf...
+ */
+static void _cli_cmd_stprivatekey_event(char *string)
+{
+    char arg[80];
+    if (ad2_copy_nth_arg(arg, string, sizeof(arg), 1) >= 0) {
+        nvs_handle my_handle;
+        esp_err_t err;
+        err = nvs_open_from_partition("stnv", "stdk", NVS_READWRITE, &my_handle);
+        if ( err != ESP_OK)
+            ESP_LOGE(TAG, "%s: nvs_open_from_partition err %d", __func__, err);
+        err = nvs_set_str(my_handle, "PrivateKey", arg);
+        if ( err != ESP_OK)
+            ESP_LOGE(TAG, "%s: nvs_set_str err %d", __func__, err);
+        err = nvs_commit(my_handle);
+        nvs_close(my_handle);
+    } else {
+        printf("Missing <privateKey>\n");
+    }
+}
+
+char * STSDK_SETTINGS [] = {
+  STSDK_CLEANUP,
+  STSDK_SERIAL,
+  STSDK_PUBKEY,
+  STSDK_PRIVKEY,
+  0 // EOF
+};
+
+static struct cli_command stsdk_cmd_list[] = {
+    {STSDK_CLEANUP,
+        "Cleanup NV data with reboot option\n"
+        "  Syntax: '" STSDK_CLEANUP "'\n", _cli_cmd_cleanup_event},
+    {STSDK_SERIAL,
+        "Sets the SmartThings device_info serialNumber.\n"
+        "  Syntax: '" STSDK_SERIAL "' <serialNumber>\n"
+        "  Example: '" STSDK_SERIAL "' AaBbCcDdEeFfGg...\n", _cli_cmd_stserial_event},
+    {STSDK_PUBKEY,
+        "Sets the SmartThings device_info publicKey.\n"
+        "  Syntax: '" STSDK_PUBKEY "' <publicKey>\n"
+        "  Example: '" STSDK_PUBKEY "' AaBbCcDdEeFfGg...\n", _cli_cmd_stpublickey_event},
+    {STSDK_PRIVKEY,
+        "Sets the SmartThings device_info privateKey.\n"
+        "  Syntax: '" STSDK_PRIVKEY "' <privateKey>\n"
+        "  Example: '" STSDK_PRIVKEY "' AaBbCcDdEeFfGg...\n", _cli_cmd_stprivatekey_event},
+};
+
 
 int get_switch_state(void)
 {
@@ -105,6 +228,7 @@ void cap_securitySystem_arm_away_cmd_cb(struct caps_securitySystem_data *caps_da
 {
     const char *init_value = caps_helper_securitySystem.attr_securitySystemStatus.value_disarmed;
     cap_securitySystem_data->set_securitySystemStatus_value(cap_securitySystem_data, init_value);
+    ad2_arm_away();
 }
 
 void cap_securitySystem_arm_stay_cmd_cb(struct caps_securitySystem_data *caps_data)
@@ -136,12 +260,15 @@ void cap_switch_cmd_cb(struct caps_switch_data *caps_data)
 
 
 void stsdk_init(void) {
-     unsigned char *onboarding_config = (unsigned char *) onboarding_config_start;
+    unsigned char *onboarding_config = (unsigned char *) onboarding_config_start;
     unsigned int onboarding_config_len = onboarding_config_end - onboarding_config_start;
     unsigned char *device_info = (unsigned char *) device_info_start;
     unsigned int device_info_len = device_info_end - device_info_start;
-
     int iot_err;
+
+    // register STSDK CLI commands
+    for (int i = 0; i < ARRAY_SIZE(stsdk_cmd_list); i++)
+        cli_register_command(&stsdk_cmd_list[i]);
 
     // create a iot context
     ctx = st_conn_init(onboarding_config, onboarding_config_len, device_info, device_info_len);
