+++ main/ad2_uart_cli.c
@@ -1,40 +1,47 @@
-/* ***************************************************************************
+/**
+ *  @file    ad2_uart_cli.c
+ *  @author  Sean Mathews <coder@f34r.com>
+ *  @date    02/20/2020
  *
- * Copyright 2019 Samsung Electronics All Rights Reserved.
+ *  @brief UART command line interface for direct access configuration
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  @copyright Copyright (C) 2020 Nu Tech Software Solutions, Inc.
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific
- * language governing permissions and limitations under the License.
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
- ****************************************************************************/
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
 
 #include <stdio.h>
-#include <string.h>
 #include <stdlib.h>
+#include <string.h>
 
+// esp includes
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
-
+#include "esp_system.h"
+#include "nvs_flash.h"
+#include "nvs.h"
+#include <lwip/netdb.h>
 #include "driver/uart.h"
-
-#include "iot_uart_cli.h"
 #include "esp_log.h"
+static const char *TAG = "AD2UTIL";
 
+// AlarmDecoder includes
+#include "alarmdecoder_main.h"
 #include "ad2_utils.h"
+#include "ad2_settings.h"
 
-static const char *TAG = "UART_CLI";
-
-#define UART_BUF_SIZE (20)
-
-#define PROMPT_STRING "AD2IOT # "
+#include "ad2_uart_cli.h"
 
 /**
 * to decide whether the main function is running or not by user action...
